Generalize the encryption method (#7)
Initially confcrypt was designed to work exclusively with RSA key on the local disk. While this works great for many usecases, it lacks generality in the encryption algorithm. This change replaces the encryptValue and decryptValue with MonadEncrypt and MonadDecrypt, which abstract out the encryption algorithm.

I still need to update the tests and add documentation for the two new monads, but otherwise this should provide the necessary generalization. I'm particularly interested in adding support for IO-only encryption methods like Amazon's KMS.